+++ backend/src/main/java/com/rayan/cinemaapi/service/UserService.java
@@ -1,7 +1,9 @@
 package com.rayan.cinemaapi.service;
 
 import com.rayan.cinemaapi.entity.User;
+import com.rayan.cinemaapi.exception.EntityInUseException;
 import com.rayan.cinemaapi.exception.EntityNotFoundException;
+import com.rayan.cinemaapi.repository.BookingRepository;
 import com.rayan.cinemaapi.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -14,10 +16,12 @@ import java.util.List;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final BookingRepository bookingRepository;
     private final PasswordEncoder passwordEncoder;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepository userRepository, BookingRepository bookingRepository, PasswordEncoder passwordEncoder) {
         this.userRepository = userRepository;
+        this.bookingRepository = bookingRepository;
         this.passwordEncoder = passwordEncoder;
     }
 
@@ -52,6 +56,10 @@ public class UserService {
     public void deleteUser(Long id) {
         if (!userRepository.existsById(id)) {
             throw new EntityNotFoundException("User", id);
+        }
+
+        if (bookingRepository.existsByUserId(id)) {
+            throw new EntityInUseException("User", id);
         }
 
         userRepository.deleteById(id);
